organize UIViewDSL logic in separate files, write different strategies for autolayout constraints
